feat: add save method for IngredientDAO

+++ src/main/java/com/restaurant/db/migration/restaurant.sql
@@ -24,7 +24,7 @@ CREATE TABLE IF NOT EXISTS ingredient (
 
 CREATE TABLE IF NOT EXISTS ingredient_price (
     ingredient_id bigint REFERENCES ingredient(ingredient_id),
-    unit_price bigint not null,
+    unit_price NUMERIC(10,2) not null,
     price_date timestamp not null,
     UNIQUE (ingredient_id, price_date)
 );
